+++ elements/x-wheelcolorpicker.js
@@ -55,6 +55,10 @@ export default class XWheelColorPickerElement extends HTMLElement {
       -webkit-user-select: none;
       --wheel-max-width: none;
     }
+    :host([disabled]) {
+      pointer-events: none;
+      opacity: 0.5;
+    }
     :host([hidden]) {
       display: none;
     }
@@ -231,6 +235,16 @@ export default class XWheelColorPickerElement extends HTMLElement {
   // @readOnly
   get computedSize() {
     return this.hasAttribute("computedsize") ? this.getAttribute("computedsize") : "medium";
+  }
+
+  // @type boolean
+  // @default false
+  // @attribute
+  get disabled() {
+    return this.hasAttribute("disabled");
+  }
+  set disabled(disabled) {
+    disabled ? this.setAttribute("disabled", "") : this.removeAttribute("disabled");
   }
 
   #shadowRoot = null;